requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.2
zeep==4.2.1
requests==2.28.1